requirements-test.txt
pytest>=8.0
pytest-playwright>=0.5
pytest-asyncio>=0.23
httpx>=0.27
pytest-cov>=5.0
